+++ packages/react/Spinner/src/spinner.definitions.ts
@@ -1,0 +1,5 @@
+export const sizes = {
+  large: 32,
+  medium: 24,
+  small: 16,
+};