+++ packages/styles/src/packages/spinner/index.ts
@@ -1,0 +1,8 @@
+import { style } from "./spinner.style.css";
+import { sprinkle, properties } from "./spinner.sprinkle.css";
+
+export const spinner = {
+  style,
+  sprinkle,
+  properties,
+};